refactor(contact-center): consolidate consult transfer list policy
